chore(*): introduce `eslint-plugin-mark` (#238)
This pull request includes updates to the `README.md` file and
configuration changes to the ESLint and package management files. The
most important changes are the consistent use of single quotes in the
`README.md`, the addition of the `eslint-plugin-mark` plugin, and
updates to the ESLint and Prettier configurations.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L104-R124):
Changed double quotes to single quotes for consistency in all sections.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L104-R124)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L140-R144)
[[3]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L211-R218)

Configuration updates:

*
[`eslint.config.mjs`](diffhunk://#diff-9601a8f6c734c2001be34a2361f76946d19a39a709b5e8c624a2a5a0aade05f2R3-R15):
Added the `eslint-plugin-mark` plugin and updated ESLint configurations
to include `bananass.configs.tsx.react` and
`mark.configs.recommendedGfm`.
[[1]](diffhunk://#diff-9601a8f6c734c2001be34a2361f76946d19a39a709b5e8c624a2a5a0aade05f2R3-R15)
[[2]](diffhunk://#diff-9601a8f6c734c2001be34a2361f76946d19a39a709b5e8c624a2a5a0aade05f2L19-L20)
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L19-R22):
Updated Prettier commands to include the `--ignore-unknown` flag and
added the `eslint-plugin-mark` dependency.
[[1]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L19-R22)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R48)