chore(*): cleanup dependencies (#286)
This pull request includes updates to the `package.json` file, primarily
focusing on dependency management. The most significant changes involve
the reorganization and relocation of the `dotenv` package and the
removal of the `@babel/cli` package.

### Dependency management updates:

* Removed the `dotenv` package from the `dependencies` section and added
it to the `devDependencies` section, reflecting its use as a
development-only tool. (`package.json`,
[[1]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L28)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L39-R44)
* Removed the `@babel/cli` package from the `devDependencies` section,
likely because it is no longer needed in the project. (`package.json`,
[package.jsonL39-R44](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L39-R44))