refactor(*): drop `react-router-dom` (#334)
This pull request simplifies the application routing setup, removes an
unused dependency, and adjusts environment variable handling in the
deployment workflow. The most important changes are grouped below:

### Simplification of application routing:
*
[`src/App.jsx`](diffhunk://#diff-d274a54187c91ba0f532df2a9e194e27ab50e988f5e4c33f5a7893918320c661L2-R10):
Removed the `BrowserRouter`, `Routes`, and `Route` components from the
application, simplifying the routing logic to directly render the
`Start` component.

### Dependency cleanup:
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L33):
Removed the `react-router-dom` dependency, as it is no longer used after
simplifying the routing logic.

### Deployment workflow adjustment:
*
[`.github/workflows/deployment.yml`](diffhunk://#diff-5a17873aec4eae6b52b00959d8f9e17672912858f63181d39de8c3a713e90018L38):
Removed the setting of the `PUBLIC_URL` environment variable in the
deployment workflow, likely because it is no longer needed in the
simplified routing logic.